feat(playground): serve page-builder pages at their own urls
Nothing rendered the page builder's output publicly. The blocks route reads
`block-pages`, whose document is authored in code, and the `blocks()` field
otherwise sat on singles, which have no public route. So a page built in the
editor could be published and then reached by nothing.

Adds one route over the `pages` collection the plugin already contributes. A
required catch-all rather than an optional one, because the app already serves
`/` from its own page and two routes claiming one path is a build error.

The reader and the site's breakpoints move to a shared module. Both public
routes need them, and two answers to "what are this site's breakpoints" is how
a page's own sheet and the shared site sheet come to emit the same tier under
different at-rules, each internally consistent and the disagreement invisible.

The blocks route's docblock claimed the two renderers store different,
non-interchangeable documents. They do not: `blocks()` declares `storage:
"json"` exactly as `json()` does, and one `BlockDocument` is defined in the
repo. It now records the real difference, which is that a json field has no
editor attached, so that collection covers the code-authored path.

The plugin's GUIDE and README told users to import `PageRenderer` from a
`/render` subpath that is not in the package's export map, and documented
`pageBuilderField` and `withPageBuilder`, neither of which exists. Five of the
eight block names the README listed were wrong. Both now describe the shipped
API, and the guide's editor-switch section goes with the `editorMode` column it
described.

No changeset: no published code changes.
